// types.ts
export interface ContactInfo {
  email1: string;
  email2: string;
  location: string;
  phone: string;
  note: string;
}
// ...
export type ApiResponse<T> =
  | { success: true; data: T; message?: string; token?: string; }
  | { success: false; message: string; };